+++ tunnels/AuthenticationServer/include/AuthenticationServer/structure.h
@@ -127,7 +127,7 @@ WW_EXPORT void         authenticationserverTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *authenticationserverTunnelCreate(node_t *node);
 WW_EXPORT api_result_t authenticationserverTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void authenticationserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void authenticationserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void authenticationserverTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void authenticationserverTunnelOnPrepair(tunnel_t *t);
 void authenticationserverTunnelOnStart(tunnel_t *t);
