+++ tunnels/EncryptionServer/include/EncryptionServer/structure.h
@@ -43,7 +43,7 @@ WW_EXPORT void         encryptionserverTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *encryptionserverTunnelCreate(node_t *node);
 WW_EXPORT api_result_t encryptionserverTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void encryptionserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void encryptionserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void encryptionserverTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void encryptionserverTunnelOnPrepair(tunnel_t *t);
 void encryptionserverTunnelOnStart(tunnel_t *t);
