+++ tunnels/ObfuscatorClient/include/ObfuscatorClient/structure.h
@@ -39,7 +39,7 @@ WW_EXPORT void         obfuscatorclientTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *obfuscatorclientTunnelCreate(node_t *node);
 WW_EXPORT api_result_t obfuscatorclientTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void obfuscatorclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void obfuscatorclientTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void obfuscatorclientTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void obfuscatorclientTunnelOnPrepair(tunnel_t *t);
 void obfuscatorclientTunnelOnStart(tunnel_t *t);
