+++ tunnels/ObfuscatorServer/include/ObfuscatorServer/structure.h
@@ -39,7 +39,7 @@ WW_EXPORT void         obfuscatorserverTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *obfuscatorserverTunnelCreate(node_t *node);
 WW_EXPORT api_result_t obfuscatorserverTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void obfuscatorserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void obfuscatorserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void obfuscatorserverTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void obfuscatorserverTunnelOnPrepair(tunnel_t *t);
 void obfuscatorserverTunnelOnStart(tunnel_t *t);
