+++ tunnels/PacketsToConnection/instance/index.c
@@ -2,7 +2,7 @@
 
 #include "loggers/network_logger.h"
 
-void ptcTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset)
+void ptcTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset)
 {
     // using tunnel / adapter default handle for this action
     discard t;
