+++ tunnels/RealityServer/include/RealityServer/structure.h
@@ -188,7 +188,7 @@ WW_EXPORT void         realityserverTunnelDestroy(tunnel_t *t);
 WW_EXPORT tunnel_t    *realityserverTunnelCreate(node_t *node);
 WW_EXPORT api_result_t realityserverTunnelApi(tunnel_t *instance, sbuf_t *message);
 
-void realityserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset);
+void realityserverTunnelOnIndex(tunnel_t *t, uint16_t index, uint32_t *mem_offset);
 void realityserverTunnelOnChain(tunnel_t *t, tunnel_chain_t *chain);
 void realityserverTunnelOnPrepair(tunnel_t *t);
 void realityserverTunnelOnStart(tunnel_t *t);
